remove op_listener
