+++ proposer/succinct/src/db/mod.rs
@@ -1,0 +1,6 @@
+mod client;
+mod types;
+
+// Re-export everything to maintain the current API
+pub use client::*;
+pub use types::*;
